requirements.txt
requests==2.31.0
fastapi
uvicorn==0.16.0
python-dotenv==0.19.2
motor == 3.7.0
python-telegram-bot== 21.10.0
playwright==1.50.0
aioredis==2.0.0
pydantic
pydantic-core
greenlet==3.1.1
pytz
bcrypt == 4.2.1